Change Domain   io.github.chaehwanli https://github.com/chaehwanli/DependencyToUml/issues/2
