+++ dependency_to_uml/build.gradle.kts
@@ -1,5 +1,3 @@
-@file:Suppress("UnstableApiUsage")
-
 /*
 * This file was generated by the Gradle 'init' task.
 *
@@ -83,8 +81,8 @@ gradlePlugin {
     vcsUrl = "https://github.com/chaehwanli/DependencyToUml.git"
     plugins {
         create("dependencyToUMLPlugin") {
-            id = "com.meteroid.dep2uml.dependencyToUMLPlugin"
-            implementationClass = "com.meteroid.dep2uml.plugin.DependencyToUMLPlugin"
+            id = "io.github.chaehwanli.dep2uml.dependencyToUMLPlugin"
+            implementationClass = "io.github.chaehwanli.dep2uml.plugin.DependencyToUMLPlugin"
             displayName = "Dependency To UML Plugin"
             description =
                 "A program that creates a package relationship using uml by referring to the dependency relationship."
