+++ dependency_to_uml/src/main/kotlin/io/github/chaehwanli/dep2uml/plugin/DependencyToUMLExtension.kt
@@ -1,0 +1,25 @@
+/*
+ *
+ * Copyright ${YEAR} chaehwan.li@gmail.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *  
+ */
+
+package io.github.chaehwanli.dep2uml.plugin
+
+class DependencyToUMLExtension {
+    var outputPath: String = "build/dependency-diagram.puml"
+    var diagramStyle: String = "default" // 예: "default", "detailed"
+    var filterDependencies: List<String> = emptyList() // 필터링할 의존성 리스트
+}
