--- dependency_to_uml/src/test/kotlin/com/meteroid/dep2uml/model/DependencyConfigurationTest.kt
@@ -1,33 +1,0 @@
-/*
- *
- * Copyright 2025 Meteroid contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.meteroid.dep2uml.model
-
-import org.junit.jupiter.api.Assertions.assertEquals
-
-import org.junit.jupiter.api.Test
-
-class DependencyConfigurationTest {
-
-    @Test
-    fun valueOf() {
-        val dependencyConfiguration = DependencyConfiguration.fromConfigurationName("api")
-
-        assertEquals(DependencyType.API, dependencyConfiguration)
-    }
-}
